movies_admin/src/requirements.txt
psycopg[binary]==3.1.10
Django==4.2.5
python-dotenv==1.0.0
django-split-settings==1.2.0
django-debug-toolbar==4.2.0
